Update patch set 1
Patch Set 1:

Well, because of the 6 different "preferred-apn*.xml" files I found in that directory, with 4 of them referring to the current apn_id of either the "phone" or "pta" apns, and the other 2 with 'apn_id's of probably the same ones before another 85 apns were added to the full apns-conf.xml file... I deleted all of them, and after rebooting now have just 3, which all reference the same apn_id (preferred-apn.xml, preferred-apn-1.xml, and preferred-apn1.xml). I'll see how my phone does with that through the rest of the day.

Patch-set: 1